Make the list of services dynamic
